requirements.txt
pandas
numpy
sqlalchemy
beautifulsoup4
requests
streamlit
scikit-learn
joblib
os
pymongo
load-dotenv
lxml